Rename loader.dart to loader_ui.dart
