+++ webapp/lib/main_ui.dart
@@ -8,7 +8,7 @@ import 'dart:html';
 import 'logger.dart' as log;
 import 'data_model.dart';
 import 'snackbar.dart' as snackbar;
-import 'loader.dart' as loader;
+import 'loader_ui.dart' as loader;
 import 'authentication.dart' as auth;
 import 'firebase_tools.dart' as fbt;
